add: Restaurant type with generics Restaurants#45
