--- src/types/GeoLocation/index.types.ts
@@ -1,4 +1,0 @@
-export type Coords = {
-  lat: number
-  lng: number
-}
