+++ src/types/Restaurant/index.ts
@@ -1,0 +1,4 @@
+export type RestaurantProps<T = {} | {}> = {
+  place_id: string
+  is_liked: boolean
+} & T
